refactor variable actions
