+++ public/app/features/dashboard-scene/settings/variables/VariableEditableElement.tsx
@@ -16,10 +16,8 @@ import {
   useSceneObjectState,
 } from '@grafana/scenes';
 import { Alert, Box, Button, Combobox, Field, Input, Stack, TextArea } from '@grafana/ui';
-import { appEvents } from 'app/core/app_events';
 import { OptionsPaneCategoryDescriptor } from 'app/features/dashboard/components/PanelEditor/OptionsPaneCategoryDescriptor';
 import { OptionsPaneItemDescriptor } from 'app/features/dashboard/components/PanelEditor/OptionsPaneItemDescriptor';
-import { ShowConfirmModalEvent } from 'app/types/events';
 
 import { DashboardScene } from '../../scene/DashboardScene';
 import { useSidebarInputAutoFocus } from '../../scene/layouts-shared/utils';
@@ -36,10 +34,10 @@ import {
   restoreUnshadowedPredefinedVariables,
   validateVariableName,
 } from '../../settings/variables/utils';
+import { confirmDeleteVariable, deleteVariable, duplicateVariable } from '../../sidebar/dashboard/variableListActions';
 import { dashboardEditActions } from '../../sidebar/shared';
 import { dashboardSceneGraph } from '../../utils/dashboardSceneGraph';
 import { getTopPlacementLabel } from '../../utils/getTopPlacementLabel';
-import { DashboardInteractions } from '../../utils/interactions';
 
 import { openChangeVariableTypePane } from './VariableTypeSelectionPane';
 import { useVariableSelectionOptionsCategory } from './useVariableSelectionOptionsCategory';
@@ -156,49 +154,15 @@ export class VariableEditableElement implements EditableDashboardElement, BulkAc
   }
 
   public onDuplicate() {
-    const set = this.variable.parent!;
-    if (!(set instanceof SceneVariableSet)) {
-      return;
-    }
-
-    dashboardEditActions.addVariable({
-      source: set,
-      addedObject: this.variable.clone({
-        key: undefined,
-        name: `${this.variable.state.name}_copy${set.state.variables.length}`,
-      }),
-    });
-    DashboardInteractions.variableActionButtonClicked('duplicate', { type: this.variable.state.type });
+    duplicateVariable(this.variable);
   }
 
   public onConfirmDelete() {
-    const name = this.variable.state.name;
-    appEvents.publish(
-      new ShowConfirmModalEvent({
-        title: t('dashboard-scene.variable-editable-element.delete-title', 'Delete variable'),
-        text: t('dashboard-scene.variable-editable-element.delete-text', 'Are you sure you want to delete: {{name}}?', {
-          name,
-        }),
-        yesText: t('dashboard-scene.variable-editable-element.delete-confirm', 'Delete variable'),
-        onConfirm: () => {
-          this.onDelete();
-        },
-      })
-    );
+    confirmDeleteVariable(this.variable);
   }
 
   public onDelete() {
-    const set = this.variable.parent!;
-    if (!(set instanceof SceneVariableSet)) {
-      return;
-    }
-
-    DashboardInteractions.variableActionButtonClicked('delete', { type: this.variable.state.type });
-
-    dashboardEditActions.removeVariable({
-      source: set,
-      removedObject: this.variable,
-    });
+    deleteVariable(this.variable);
   }
 
   public onChangeName(name: string) {
